AI-Invoice/invoice-ai/requirements.txt
python-dotenv==1.0.1
openai==1.55.3
openpyxl==3.1.5
jsonschema==4.23.0
